refactor(workflow): share llm provider resolution
