+++ lib/crates/fabro-workflow/src/handler/llm/routing.rs
@@ -1,4 +1,5 @@
 use fabro_graphviz::graph::{self, Node};
+use fabro_model::{AgentProfileKind, Catalog, ProviderId};
 use fabro_types::LlmBackend;
 
 use super::cli::is_cli_only_model;
@@ -41,6 +42,54 @@ pub(crate) fn node_needs_api_backend(node: &Node) -> bool {
         }
         _ => matches!(select_run_backend(node), Ok(LlmBackend::Api)),
     }
+}
+
+#[derive(Clone)]
+pub(super) struct ProviderContext {
+    pub(super) provider_id:  ProviderId,
+    pub(super) profile_kind: AgentProfileKind,
+}
+
+pub(super) fn default_profile_kind(
+    catalog: &Catalog,
+    provider_id: &ProviderId,
+) -> AgentProfileKind {
+    catalog
+        .provider(provider_id)
+        .unwrap_or_else(|| panic!("Provider \"{provider_id}\" is not configured"))
+        .adapter
+        .metadata()
+        .default_profile
+}
+
+pub(super) fn resolve_provider_context(
+    catalog: &Catalog,
+    default_provider_id: &ProviderId,
+    model: &str,
+    provider_attr: Option<&str>,
+) -> Result<ProviderContext, Error> {
+    let provider_id = if let Some(provider) = provider_attr {
+        let requested = ProviderId::from(provider);
+        catalog
+            .provider(&requested)
+            .ok_or_else(|| {
+                Error::Precondition(format!("Provider \"{provider}\" is not configured"))
+            })?
+            .id
+            .clone()
+    } else if let Some(model) = catalog.get(model) {
+        model.provider.clone()
+    } else {
+        default_provider_id.clone()
+    };
+
+    let provider = catalog.provider(&provider_id).ok_or_else(|| {
+        Error::Precondition(format!("Provider \"{provider_id}\" is not configured"))
+    })?;
+    Ok(ProviderContext {
+        provider_id:  provider.id.clone(),
+        profile_kind: provider.adapter.metadata().default_profile,
+    })
 }
 
 fn unsupported_backend_error(raw: &str) -> Error {
